backend/requirements.txt
flask
werkzeug
gunicorn
psycopg2-binary
flask-cors
